Update the Asset query API
Previously, we do not have the possibility to return the general
info about the registered orders for the given Asset. Now, we the
user can request the API to return the general info. This general
info has two fields: totalQuantity and weightedPrice. The former
represents how much asset "papers" you have, with the latter
meaning your average price based on the registered orders.