Update the API which query an asset from our database
This commit improve how we fetch this information:
- The REST API returns only field that are not empty.
- The OrderInfo has now three fields: TotalQuantity,
WeightedAdjPrice and WeightedAveragePrice. The first two was
already implemented, with the latter representing the average
price from the Asset. This average price only considers the buy
orders for its measurement. It is used to measure the average
price for the Asset which we declare for the Federal Government.