Reorganize project folders
This commit improves the project organization. Now, we have three
folders: database, handlers and router. The database folder has
the queries for each table in our PostgreSQL database and the file
that will be used in the future to connect with our database. In
relation to router, this folder has our REST API. Lastly, handlers
has the functions to be executed for each REST API.