Implement Asset verification before order and asset creation
Previously, the backend does not search if the given asset exist
or not. Now, the backend verifies if the given asset exist before
its creation in the database. Such action is done in the order and
asset handlers. Furthermore, to enable this research this commit
implements the Company Profile and Overview based on the Finnhub
and Alpha Vantage API, respectively.