Update from the pgxpool to pgx to access the postgreSQL database
In this commit, we change the library to access the postgreSQL
database because there isn't an advanced library for database
mocking considering the pgxpool. Meanwhile, for the pgx, there is
a more advanced package that has the purpose to mock our database.

For future implementations, we can go back for the pgxpool that is
a more efficient package for database request and implements an
original package that mock this database via pgxpool.