Implement PostgreSQL users table
This commit implements the users table on the PostgreSQL database
to control in the future for example which order, earning and
assets each user has. Furthermore, this commit updates the handlers
from intermediary REST API for Firebase, with the purpose to
include the user information update in our PostgreSQL database.