Include authentication in the asset.go functions
This commit is the first commit from this branch with the purpose
to include writing and reading in our database for data related
with the sent token. Specifically in this commit, we implement such
authentication in our file responsible to execute function related
to the Asset table in our postgresql.