Update order API to include authentication
This commit introduces the authentication and user verification
before requesting order creation, update or delete. Now, the
backend will populate the user_uid column in the order table in
our PostgreSQL database to specify that the order is from the user
with that given UID. The UID is defined based on the user UID that
is associated with the Bearer token that goes on the request.
Furthermore, to update and delete order via API now the request
is only able to execute something in our database when the orderId
matches with the associated user UID that was defined when the user
created for the firts time that given order.