Implement driver for postgresql following Clean Arch
This commmit is the first step towards the Clean Architecture.
Specifically in this commit, we implement the concepts of entity,
repository to write in any possible database configuration.

In this actual state, the application does not run because this
change broke the Fiber application. The next step is to refactor
the fiber application to follow the Clean Architecture principles.