Implement the first application logic following Clean Arch
This commit is the initial step to update the handlers from the
API. Following the Clean Architecture, all the application logic
and data verification should be done in this layer. Hence, this is
the first commit focusing the application logic to enable the
Asset creation. The advantages to use this technique is because
the application logic does not know what is the database used and
port for the outside world.