Initial setup using dockerfile
This commit represents the initial setup to use docker in our
Stockfy Application. In this setup, we have two dockerfiles. One
to create our database for production (Dockerfile) and the other
is responsible to build and run the binary file from the golang
application (Dockerfile.multistage).

The next step is to create a docker for integration tests, and to
create a docker compose for automation.