Implement go routine for asset info and price (#11)
* Implement go routine for asset info and price

* Mock Finnhub and Alpha API for integration test

* Update asset integration test

This commit updates the TestFiberHandlersIntegrationTestGetAsset
to include the case when the user request that the asset information
returns with its correspondent price.

* Create an entity funtion to conver str to time.Time

* Implement withPrice URL query for assetType handler

* Include withPrice query in the asset type handler

This commit implements a new query called AssetPrice. The purpose
for this variable is to define if the request needs to return or not
the current price for each asset.

* Improve RESTful API to get asset info and price

This commit removes the finnhub and alpha vantage endpoints. Now,
we have only one endpoint to get the asset info and price.